Move LoggerConfig class to the vantiqservicesdk module
Doing multiple files/directories is more complicated and I don't want to
sort that right now
