+++ pythonsdk/build.gradle
@@ -13,7 +13,7 @@ ext {
         pypi_repo = 'testpypi'
     }
     if (!project.rootProject.hasProperty('vantiqServiceSdkVersion')) {
-        vantiqServiceSdkVersion = '0.0.5'
+        vantiqServiceSdkVersion = '0.0.6'
     }
     
     // See if we are using virtualenvwrapper and if so put the virtualenv in the WORKON_HOME
